rename ServiceModule to DatabaseModule and move database initialization to it
